Risk of bias bar plot

To plot risk of bias result.

RoB_plot(
  study.CM,
  rob.ind,
  lab.size,
  font.family = c("Helvetica", "Times New Roman")
)

Arguments

study.CM
contribution matrix

rob.ind
risk of bias or indirectness

lab.size
label size

font.family
"Helvetica" or "Times New Roman"

Value

ggplot object

Examples

LDT1 <- read.csv(system.file("extdata", "HR_SH_D.csv", package = "net.meta"))
trt1 <- read.table(system.file("extdata", "HR_SH_D.txt", package = "net.meta"),
    header=TRUE,quote = '"', stringsAsFactors=FALSE)

trt1$description <- factor(trt1$description, trt1$description)
LDT1$study <- factor(LDT1$study, unique(LDT1$study))

nmt1 <- model_netmeta(long.data = LDT1,
treatment=LDT1$treatment,
id.treatments = trt1,
reference = "A",
outcome = "HR")

name1 <- NULL
for(i in 1:(length(trt1$id)-1)){
  name1 <- c(name1, paste0(trt1$id[i], ":", trt1$id[-(1:i)]))
}

study.CM1 <- studyCM_matrix(name1, nmt1)

study.assess1 <- read.csv(system.file("extdata", "HR_SH_A.csv", package = "net.meta"))

p<-RoB_plot(study.CM1, study.assess1$ROB, lab.size=10)

#> Warning: Vectorized input to `element_text()` is not officially supported.
#> Results may be unexpected or may change in future versions of ggplot2.

p<-RoB_plot(study.CM1, study.assess1$IND, lab.size=10)

#> Warning: Vectorized input to `element_text()` is not officially supported.
#> Results may be unexpected or may change in future versions of ggplot2.