Forest-table plot of net-meta model for MD, HR, and RR

To give the forest plot with table of bayesian results from model_gemtc or combined forest plot of bayesian results, frequentist results, and prediction interval from model_gemtc and model_netmeta.

forest_table_plot(
  type = c("bayesian", "all"),
  bmt,
  nmt = NULL,
  digits = 2,
  x1,
  x2,
  x.lab,
  title.size,
  text.size,
  point.size = 4,
  labels = NULL,
  plot.scale = 0.9,
  font.family = c("Helvetica", "Times New Roman")
)

Arguments

type
the type of plots, bayesian-bayesian result; all-bayesian, freq, and prediction interval

bmt
bayesian net-meta result from model_gemtc

nmt
frequentist net-meta result from model_nemeta

digits
digits of the results

x1
the left xlim

x2
the right xlim

x.lab
label in the x-axis

title.size
title size

text.size
font size

point.size
point size

labels
label of the whole plot

plot.scale
the scaled size of the whole plot

font.family
"Helvetica" or "Times New Roman"

Value

ggplot object

Examples

LDT1 <- read.csv(system.file("extdata", "HR_SH_D.csv", package = "net.meta"))
trt1 <- read.table(system.file("extdata", "HR_SH_D.txt", package = "net.meta"),
    header=TRUE,quote = '"', stringsAsFactors=FALSE)

trt1$description <- factor(trt1$description, trt1$description)
LDT1$study <- factor(LDT1$study, unique(LDT1$study))

bmt1 <- model_gemtc(
long.data=LDT1,
id.treatments=trt1,
reference="A",
outcome="HR",
mtc.n.adapt = 5000, mtc.n.iter = 10000, mtc.thin = 20)

#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 34
#>    Unobserved stochastic nodes: 43
#>    Total graph size: 643
#> 
#> Initializing model
#> 

p<-forest_table_plot(
type="bayesian",
bmt=bmt1,
x.lab = "X caption",
x1=0,
x2=40,
title.size =12,
text.size =4,
digits = 2,
font.family="Helvetica")

nmt1 <- model_netmeta(long.data = LDT1,
                      treatment=LDT1$treatment,
                      id.treatments = trt1,
                      reference = "A",
                      outcome = "HR")

# View(nmt1)

p<-forest_table_plot(
type="all",
bmt=bmt1,
nmt=nmt1,
x.lab = "X caption",
x1=0,
x2=50,
title.size =12,
text.size =4,
digits = 2,
font.family="Times New Roman")